base_classes {S7}	R Documentation

S7 wrappers for base types

Description

The following S7 classes represent base types allowing them to be used within S7:

class_logical

class_integer

class_double

class_complex

class_character

class_raw

class_list

class_expression

class_name

class_call

class_function

class_environment (can only be used for properties)

We also include three union types to model numerics, atomics, and vectors respectively:

class_numeric is a union of class_integer and class_double.

class_atomic is a union of class_logical, class_numeric, class_complex, class_character, and class_raw.

class_vector is a union of class_atomic, class_list, and class_expression.

class_language is a union of class_name and class_call.

Usage

class_logical

class_integer

class_double

class_complex

class_character

class_raw

class_list

class_expression

class_name

class_call

class_function

class_environment

class_numeric

class_atomic

class_vector

class_language

Value

S7 classes wrapping around common base types and S3 classes.

Examples

class_integer
class_numeric
class_factor

[Package S7 version 0.2.1 Index]