method_explain {S7}	R Documentation

Explain method dispatch

Description

method_explain() shows all possible methods that a call to a generic might use, which ones exist, and which one will actually be called.

Note that method dispatch uses a string representation of each class in the class hierarchy. Each class system uses a slightly different convention to avoid ambiguity.

S7: pkg::class or class

S4: S4/pkg::class or S4/class

S3: class

Usage

method_explain(generic, class = NULL, object = NULL)

Arguments

generic
A generic function, i.e. an S7 generic, an external generic, an S3 generic, or an S4 generic.

class, object
Perform introspection either with a class (processed with as_class()) or a concrete object. If generic uses multiple dispatch then both object and class must be a list of classes/objects.

Value

Nothing; this function is called for it's side effects.

Examples

Foo1 <- new_class("Foo1")
Foo2 <- new_class("Foo2", Foo1)

add <- new_generic("add", c("x", "y"))
method(add, list(Foo2, Foo1)) <- function(x, y) c(2, 1)
method(add, list(Foo1, Foo1)) <- function(x, y) c(1, 1)

method_explain(add, list(Foo2, Foo2))

[Package S7 version 0.2.1 Index]